+++ example/lib/main.dart
@@ -156,7 +156,7 @@ class _MyHomePageState extends State<MyHomePage> with DynamicRoutesInitiator {
             dynamicRoutesInitiator.initializeRoutes(_widgets,
                 lastPageCallback: (newContext) {
               Navigator.popUntil(newContext, (route) => route.isFirst);
-            }, errorBuilder: (context, errorDetails) {});
+            });
 
             dynamicRoutesInitiator.pushFirst(context).then((_) {
               setState(() {});