+++ CHANGELOG.md
@@ -11,6 +11,6 @@ Name change and docs edit
 - Some parts of the docs and the variable names were still referencing the old "StackedRoutes" name.
 - Added a method to pop for a number of pages between 0 to pages.length.
 - DynamicRoutesNavigator.pop now conforms to the Navigator.pop interface.
-- Added a method to pop to a specific page.
+- Added popFor, pushFor, and popRange.
 - Can now control the transition animation.
 - The initiator example in the docs was missing a step.