+++ CHANGELOG.md
@@ -8,9 +8,15 @@ Name change and docs edit
 
 ## 1.1.0
 
-- Some parts of the docs and the variable names were still referencing the old "StackedRoutes" name.
+### New features.
+
 - Added a method to pop for a number of pages between 0 to pages.length.
 - DynamicRoutesNavigator.pop now conforms to the Navigator.pop interface.
 - Added popFor, pushFor, and popRange.
 - Can now control the transition animation.
+- Sub-routing now possible.
+
+### Docs-related
+
 - The initiator example in the docs was missing a step.
+- Some parts of the docs and the variable names were still referencing the old "StackedRoutes" name.